+++ deploy/010-deploy-reward-distributors.ts
@@ -65,7 +65,10 @@ const func: DeployFunction = async function (hre: HardhatRuntimeEnvironment) {
       });
 
       const rewardsDistributor = await ethers.getContract(contractName);
-      if ((await rewardsDistributor.owner()) === deployer) {
+      if (
+        (await rewardsDistributor.owner()) === deployer &&
+        (await rewardsDistributor.pendingOwner()) === ethers.constants.AddressZero
+      ) {
         await rewardsDistributor.transferOwnership(proxyOwnerAddress);
       }
     }